InMemoryStorage extracted
